requirements.txt
requests>=2.0
psycopg2-binary>=2.9
Flask>=2.2
python-dotenv>=0.21
